fixes js docs
